refactor(perf): extract business module, split snapshot into business/snapshot.py
Move snapshot CRUD from a standalone cleanup_snapshots.py into a
dedicated business/ sub-package:

  tests/perf/business/
    __init__.py     package init + doc
    snapshot.py     list_snaps / delete_snaps / auto_cleanup_if_enabled

__main__.py imports updated from `cleanup_snapshots` to
`business.snapshot`. The public API (list_snaps, delete_snaps,
auto_cleanup_if_enabled) is unchanged — only the module path moved.

This gives a clean seam: framework/ is the execution engine, reporting/
handles data assembly, plugins/ wraps HTML generation, and business/
holds domain operations (snapshot CRUD, future: template provisioning,
sandbox lifecycle).
